refactor(index): drop `Spinner.js` in favor of `index.js`
